+++ src/meteorological_preprocessor/aws/tar_txt_index_main_function.sh
@@ -44,8 +44,7 @@ function handler () {
       echo ${destination_tar_index_directory}/tar/${extension_type}/${tar_file} > ${work_directory}/tar/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt
       cp /dev/null ${work_directory}/err_log.tmp
       set +e
-#      timeout -k 3 30 rclone copy --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/tar/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt --log-file ${work_directory}/err_log.tmp --local-no-set-modtime --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --azureblob-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket} ${work_directory}/${sub_directory}/processed/${extension_type}
-      timeout -k 3 30 rclone copy --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/tar/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt --log-file ${work_directory}/err_log.tmp --local-no-set-modtime --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket} ${work_directory}/${sub_directory}/processed/${extension_type}
+      timeout -k 3 30 rclone copy --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/tar/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt --local-no-set-modtime --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --azureblob-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket} ${work_directory}/${sub_directory}/processed/${extension_type}
       exit_code=$?
       set -e
       if test ${exit_code} -ne 0; then
@@ -88,8 +87,7 @@ function handler () {
       echo ${destination_tar_index_directory}/index/${extension_type}/${index_file} > ${work_directory}/index/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt
       cp /dev/null ${work_directory}/err_log.tmp
       set +e
-#      timeout -k 3 30 rclone copy --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/index/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt --local-no-set-modtime --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --azureblob-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket} ${work_directory}/${sub_directory}/${center_id}/${extension_type}/${destination_rclone_remote_bucket_directory}
-      timeout -k 3 30 rclone copy --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/index/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt --local-no-set-modtime --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket} ${work_directory}/${sub_directory}/${center_id}/${extension_type}/${destination_rclone_remote_bucket_directory}
+      timeout -k 3 30 rclone copy --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/index/${extension_type}/${destination_rclone_remote_bucket_directory}/raw.txt --local-no-set-modtime --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --azureblob-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket} ${work_directory}/${sub_directory}/${center_id}/${extension_type}/${destination_rclone_remote_bucket_directory}
       set -e
       if test ${exit_code} -ne 0; then
         cat ${work_directory}/err_log.tmp >&2
@@ -136,17 +134,27 @@ function handler () {
         echo "ERROR: ${function} can not pub a file of ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/tar/${extension_type}/${now}.tar." >&2
         rc=${exit_code}
       fi
-      for tar_file in `timeout -k 3 30 rclone lsf --config rclone.conf --contimeout 8s --low-level-retries 3 --max-depth 1 --no-traverse --quiet --retries 3 --stats 0 --timeout 8s ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/tar/${extension_type}/ | grep -E '^[0-9]+\.tar' | head -n -2`; do
+      cp /dev/null ${work_directory}/delete.txt
+      cp /dev/null ${work_directory}/err_log.tmp
+      set +e
+      timeout -k 3 30 rclone lsf --config rclone.conf --contimeout 8s --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --max-depth 1 --no-traverse --quiet --retries 3 --stats 0 --timeout 8s ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/tar/${extension_type}/ | grep -E '^[0-9]+\.tar' | sort -u | head -n -2 | sed -e "s|^|${destination_tar_index_directory}/tar/${extension_type}/|g"> ${work_directory}/delete.txt
+      exit_code=$?
+      set -e
+      if test ${exit_code} -ne 0; then
+        cat ${work_directory}/err_log.tmp >&2
+        echo "INFO: ${function} can not get a list of ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/tar/${extension_type}/" >&2
+      fi
+      if test -s ${work_directory}/delete.txt; then
         cp /dev/null ${work_directory}/err_log.tmp
         set +e
-        timeout -k 3 30 rclone delete --config rclone.conf --contimeout 8s --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --stats 0 --timeout 8s ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/tar/${extension_type}/${tar_file}
+        timeout -k 3 30 rclone delete --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/delete.txt --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --azureblob-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket}
         exit_code=$?
         set -e
         if test ${exit_code} -ne 0; then
           cat ${work_directory}/err_log.tmp >&2
-          echo "INFO: ${function} can not delete a file of ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/tar/${extension_type}/${tar_file}" >&2
+          echo "INFO: ${function} can not delete old files in ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/tar/${extension_type}/" >&2
         fi
-      done
+      fi
     fi
     if test -s ${work_directory}/${sub_directory}/${center_id}/${extension_type}/${destination_rclone_remote_bucket_directory}/4PubSub_index.txt; then
       cp /dev/null ${work_directory}/err_log.tmp
@@ -159,17 +167,27 @@ function handler () {
         echo "ERROR: ${function} can not pub a file of ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/index/${extension_type}/${now}.txt." >&2
         rc=${exit_code}
       fi
-      for index_file in `timeout -k 3 30 rclone lsf --config rclone.conf --contimeout 8s --low-level-retries 3 --max-depth 1 --no-traverse --quiet --retries 3 --stats 0 --timeout 8s ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/index/${extension_type}/ | grep -E '^[0-9]+\.txt' | head -n -2`; do
+      cp /dev/null ${work_directory}/delete.txt
+      cp /dev/null ${work_directory}/err_log.tmp
+      set +e
+      timeout -k 3 30 rclone lsf --config rclone.conf --contimeout 8s --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --max-depth 1 --no-traverse --quiet --retries 3 --stats 0 --timeout 8s ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/index/${extension_type}/ | grep -E '^[0-9]+\.txt' | sort -u | head -n -2 | sed -e "s|^|${destination_tar_index_directory}/index/${extension_type}/|g" > ${work_directory}/delete.txt
+      exit_code=$?
+      set -e
+      if test ${exit_code} -ne 0; then
+        cat ${work_directory}/err_log.tmp >&2
+        echo "INFO: ${function} can not get a list of ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/index/${extension_type}/" >&2
+      fi
+      if test -s  ${work_directory}/delete.txt; then
         cp /dev/null ${work_directory}/err_log.tmp
         set +e
-        timeout -k 3 30 rclone delete --config rclone.conf --contimeout 8s --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --stats 0 --timeout 8s ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/index/${extension_type}/${index_file}
+        timeout -k 3 30 rclone delete --checksum --config rclone.conf --contimeout 8s --files-from-raw ${work_directory}/delete.txt --log-file ${work_directory}/err_log.tmp --low-level-retries 3 --no-traverse --quiet --retries 3 --s3-no-check-bucket --s3-no-head --s3-no-head-object --azureblob-no-head-object --stats 0 --timeout 8s ${destination_rclone_remote_bucket}
         exit_code=$?
         set -e
         if test ${exit_code} -ne 0; then
           cat ${work_directory}/err_log.tmp >&2
-          echo "INFO: ${function} can not delete a file of ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/index/${extension_type}/${index_file}" >&2
+          echo "INFO: ${function} can not delete old files in ${destination_rclone_remote_bucket}/${destination_tar_index_directory}/index/${extension_type}/" >&2
         fi
-      done
+      fi
     fi
   done
   return ${rc}
